backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pinecone==5.0.1
openai==1.45.0
PyJWT==2.9.0
bcrypt==4.2.0
pypdf==4.3.1
python-docx==1.1.2
python-dotenv==1.0.1
sentence-transformers==3.1.1
